# share-this/api-spec.yaml
openapi: 3.0.3
info:
  title: JADS Platform API
  version: 4.0.0
  description: >
    Joint Airspace Drone System — India's sovereign airspace management
    and forensic audit platform. Handles ICAO-compliant flight plan filing
    for manned aircraft and forensic-grade drone mission telemetry.
  contact:
    name: JADS Platform Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api
    description: Local development server

# ── Global header requirement ───────────────────────────────────────────────
# All requests must include X-JADS-Version: 4.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        User JWT (signed with JWT_SECRET, 8h expiry) or
        Admin JWT (signed with ADMIN_JWT_SECRET, 2h expiry).
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
...
security:
  - bearerAuth: []
paths:
...
  /flight-plans:
    post:
      tags: [Flight Plans]
      summary: File a manned aircraft flight plan
      description: >
        Validates ICAO OFPL fields, builds AFTN message (FPL format),
        generates AFTN addressees, transmits via AFTN gateway.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [aircraftId, adep, ades, eobt, cruisingSpeed, cruisingLevel, route]
              properties:
                aircraftId: { type: string, description: 'Aircraft registration (callsign)' }
                adep: { type: string, description: 'Departure ICAO code (4 chars)' }
                ades: { type: string, description: 'Destination ICAO code (4 chars)' }
                eobt: { type: string, description: 'Estimated off-block time (ISO 8601 or DDHHmm)' }
                cruisingSpeed: { type: string, description: 'ICAO speed (e.g. N0480, K0890, M080)' }
                cruisingLevel: { type: string, description: 'Flight level (e.g. F350, VFR)' }
                route: { type: string, description: 'Route string (airways or DCT)' }
                flightRules: { type: string, enum: [IFR, VFR, Y, Z] }
                flightType: { type: string }
                aircraftType: { type: string }
                wakeTurbulence: { type: string, enum: [L, M, H, J] }
                equipment: { type: string }
                surveillance: { type: string }
                altn1: { type: string }
                altn2: { type: string }
                endurance: { type: string }
                personsOnBoard: { type: integer }
                item18: { type: string, description: 'Other info (DOF, REG, OPR, etc.)' }
      responses:
        '201':
          description: Flight plan filed successfully
        '422':
          description: OFPL validation failed

    get:
      tags: [Flight Plans]
      summary: List filed flight plans for current user
      responses:
        '200':
          description: List of flight plans (max 50, newest first)
  /flight-plans/{id}:
    get:
      tags: [Flight Plans]
      summary: Get flight plan details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Full flight plan with ADC/FIC references
        '404':
          description: Flight plan not found

    put:
      tags: [Flight Plans]
      summary: Amend a filed flight plan (before clearance only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                eobt: { type: string }
                route: { type: string }
                cruisingLevel: { type: string }
                cruisingSpeed: { type: string }
                altn1: { type: string }
                altn2: { type: string }
      responses:
        '200':
          description: Flight plan amended
        '409':
          description: Cannot edit after clearance
  /flight-plans/{id}/cancel:
    post:
      tags: [Flight Plans]
      summary: Cancel a filed flight plan (generates AFTN CNL)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200':
          description: Plan cancelled — AFTN CNL transmitted
        '409':
          description: Cannot cancel in current status
  /flight-plans/{id}/delay:
    post:
      tags: [Flight Plans]
      summary: Delay a filed flight plan (generates AFTN DLA)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newEobt, reason]
              properties:
                newEobt: { type: string, description: 'New EOBT in DDHHmm format' }
                reason: { type: string }
      responses:
        '200':
          description: Plan delayed — AFTN DLA transmitted
  /flight-plans/{id}/events:
    get:
      tags: [Flight Plans]
      summary: SSE stream for real-time clearance updates
      description: >
        Server-Sent Events stream. Pushes events as AFMLU/FIR issue
        ADC and FIC clearance numbers. Events: connected, adc_issued,
        fic_issued, clearance_rejected, keepalive (every 25s).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: SSE stream opened
          content:
            text/event-stream: {}
  /flight-plans/route-advisory:
    post:
      tags: [Flight Plans]
      summary: Get route advisory (recommended airway, FL, reporting points)
      description: >
        Advisory-only. Returns recommended airway route, flight level
        compliance (semicircular rule), reporting points, FIR crossings,
        and direct route comparison. No DB writes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adep, ades]
              properties:
                adep: { type: string }
                ades: { type: string }
                cruisingLevel: { type: string, default: VFR }
                cruisingSpeed: { type: string, default: N0240 }
      responses:
        '200':
          description: Route advisory generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  advisory:
                    type: object
                    properties:
                      hasRecommendation: { type: boolean }
                      recommended:
                        type: object
                        nullable: true
                        properties:
                          routeString: { type: string }
                          totalDistanceNm: { type: number }
                          totalEetMinutes: { type: number }
                      flightLevelAdvisory:
                        type: object
                        properties:
                          requestedLevel: { type: string }
                          isCompliant: { type: boolean }
                          recommendedLevel: { type: string }
                          direction: { type: string, enum: [EASTBOUND, WESTBOUND] }

  # ═══════════════════════════════════════════════════════════════════════════
  # AUDIT — Forensic audit & evidence
  # ═══════════════════════════════════════════════════════════════════════════
...
  /audit/flight-plans:
    get:
      tags: [Audit]
      summary: List flight plans (auditor view)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: callsign
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Paginated flight plan list
...